Server: refactor index.js for readability
