+++ index.js
@@ -9,7 +9,6 @@ require('babel/register')({
   stage: 0,
 });
 
-var config;
 var errorLog = require('./src/lib/errorLog');
 
 process.on('uncaughtException', function (err) {
@@ -48,67 +47,17 @@ var cluster = require('cluster');
 var numCPUs = process.env.PROCESSES || require('os').cpus().length;
 
 // App config
-var defaultConfig = require('./src/config');
-var config = defaultConfig();
+var config = require('./src/serverConfig')(numCPUs);
 
-// Import built-asset manifests for passing to layouts
-var jsManifest = require('./build/js/client-manifest.json');
-var cssManifest = require('./build/css/css-manifest.json');
 var servers = [];
 
 var failedProcesses = 0;
-
-// Then merge them into a single object for ease of use later
-config.manifest = {};
-config.processes = numCPUs;
-
-function parseObject (list) {
-  if (!list) { return; }
-  var obj = {};
-  var key;
-  var value;
-  var split;
-
-  list.split(';').forEach(function (l) {
-    if (l && l.indexOf('=')) {
-      var split = l.split('=');
-      obj[split[0].trim()] = split[1].trim();
-    }
-  });
-
-  return obj;
-}
-
-function parseList (list) {
-  if (!list) { return; }
-  return list.split(';');
-}
-
-config.apiHeaders = parseObject(process.env.API_HEADERS);
-config.apiPassThroughHeaders = parseList(process.env.API_PASS_THROUGH_HEADERS);
-config.actionNameSecret = process.env.ACTION_NAME_SECRET;
-
-config.experiments = parseObject(process.env.EXPERIMENTS);
-
-Object.assign(config.manifest, jsManifest, cssManifest);
 
 function start(config) {
   var server = new Server(config);
   server.start();
   return server;
 }
-
-// Private, server-only config that we don't put in config.js, which is shared
-config.liveReload = process.env.LIVERELOAD === 'true';
-config.oauth = {
-  clientId: process.env.OAUTH_CLIENT_ID || '',
-  secret: process.env.OAUTH_SECRET || '',
-
-  secretClientId: process.env.SECRET_OAUTH_CLIENT_ID || '',
-  secretSecret: process.env.SECRET_OAUTH_SECRET || '',
-};
-
-config.keys = [ process.env.SERVER_SIGNED_COOKIE_KEY || 'lambeosaurus' ];
 
 if (cluster.isMaster) {
   for (var i = 0; i < numCPUs; i++) {
